Registery of all "found" entries for AI Archetypes
Left Hand:
MODIFIER => Take current multiplier and ADD the value.
MULTIPLIER => Set value of the weight.

Right-Right-Hand: (ENEMY, ALLY, TARGET, STUPIDITY, etc...) these are the finality.

Right-Left-Hand (middle): _<FORMAT>_
These are context.

I don't know how to create new entries, but this is a registery for all existing entry I found in the ~160 Ai Archetype there in Norbyte's Search Engine. Each of these entry are unique in this file. To modify these, just create a new file and add.

Note: "+" or "-" can be added at the begining of the line and/or before the number, which means that
-MODIFIER_CONTROL_STUPIDITY -0.75 actually adds it up.